Add Course Complete

+++ server/controllers/Section.js
@@ -1,5 +1,6 @@
 const Section=require("../models/Section");
 const Course=require("../models/Course");
+const SubSection = require("../models/SubSection");
 
 exports.createSection = async (req, res) => {
   try{
@@ -46,7 +47,7 @@ exports.createSection = async (req, res) => {
 exports.updateSection = async (req, res) => {
   try{
     //Data input
-    const {sectionName, sectionId} = req.body;
+    const {sectionName, sectionId, courseId} = req.body;
 
     //Data validation
     if (!sectionName || !sectionId) {
@@ -58,15 +59,25 @@ exports.updateSection = async (req, res) => {
     //Update data
     const section=await Section.findByIdAndUpdate(sectionId, {sectionName}, {new:true});
 
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec()
+
     //Returning response
     return res.status(200).json({
-      message: "Section updated successfully",
-      data: section
+      success: true,
+      message: section,
+      data: course
     });
 
   } catch(err){
     return res.status(500).json({
-      success:true,
+      success:false,
       message:"Error while updating section",
     });
   }
@@ -75,22 +86,45 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
   try{
     //Getting ID
-    const {sectionId}=req.body;
+    const {sectionId, courseId}=req.body;
+    await Course.findByIdAndUpdate(courseId, {
+      $pull: {
+        courseContent: sectionId,
+      },
+    })
+
+    const section = await Section.findById(sectionId)
+    console.log(sectionId, courseId)
+    if (!section) {
+      return res.status(404).json({
+        success: false,
+        message: "Section not found",
+      })
+    }
 
     //Deleting ID
-    const section = await Section.findByIdAndDelete(sectionId);
+    await SubSection.deleteMany({ _id: { $in: section.subSection } })
+    await Section.findByIdAndDelete(sectionId);
 
-    //Updating Course
+    const course = await Course.findById(courseId)
+      .populate({
+        path: "courseContent",
+        populate: {
+          path: "subSection",
+        },
+      })
+      .exec()
 
     //Returning response
     return res.status(200).json({
+      success: true,
       message: "Section deleted successfully",
-      data: section
+      data: course
     });
 
   } catch(err){
     return res.status(500).json({
-      success:true,
+      success:false,
       message:"Error while deleting section",
     });
   }
